+++ cli/src/cs/api/paginate/ApiResponse.ts
@@ -1,0 +1,4 @@
+export default interface ApiResponse {
+	readonly data?: unknown;
+	readonly error?: unknown;
+}
